# web/content/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ProxyWhirl API",
    "description": "REST API for advanced proxy rotation with auto-fetching, validation, and persistence. Manage proxy pools, make proxied requests, and monitor system health through RESTful endpoints.",
    "version": "1.0.0"
  },
  "paths": {
    "/api/request": {
      "post": {
        "tags": [
          "Proxied Requests"
        ],
        "summary": "Make proxied HTTP request",
        "description": "Make an HTTP request through a rotating proxy.\n\nThis endpoint routes your HTTP request through the proxy pool, automatically\nhandling rotation and failover.\n\nSECURITY: All target URLs are validated to prevent SSRF attacks. The following\nare blocked by default:\n- Localhost and loopback addresses (127.0.0.0/8, ::1)\n- Private IP ranges (10.0.0.0/8, 172.16.0.0/12, 192.168.0.0/16)\n- Link-local addresses (169.254.0.0/16)\n- Internal domain names (.local, .internal, .lan, .corp)\n- Non-HTTP/HTTPS schemes (file://, data://, etc.)\n\nArgs:\n    request_data: Request details (URL, method, headers, body, timeout) - validated for SSRF\n    rotator: ProxyWhirl dependency injection\n    api_key: API key verification dependency\n\nReturns:\n    APIResponse with proxied response data\n\nRaises:\n    HTTPException: For various error conditions including SSRF protection",
        "operationId": "make_proxied_request_api_request_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProxiedRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_ProxiedResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/proxies": {
      "get": {
        "tags": [
          "Pool Management"
        ],
        "summary": "List all proxies",
        "description": "List all proxies in the pool with pagination and filtering.\n\nArgs:\n    page: Page number (1-indexed)\n    page_size: Number of items per page (max 100)\n    status_filter: Filter by health status (optional)\n    rotator: ProxyWhirl dependency\n    api_key: API key verification\n\nReturns:\n    Paginated list of proxy resources",
        "operationId": "list_proxies_api_proxies_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Page Size"
            }
          },
          {
            "name": "status_filter",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status Filter"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_PaginatedResponse_ProxyResource__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Pool Management"
        ],
        "summary": "Add new proxy",
        "description": "Add a new proxy to the pool.\n\nArgs:\n    proxy_data: Proxy URL and optional credentials\n    rotator: ProxyWhirl dependency\n    api_key: API key verification\n\nReturns:\n    Created proxy resource",
        "operationId": "add_proxy_api_proxies_post",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProxyRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_ProxyResource_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/rotate": {
      "get": {
        "tags": [
          "Pool Management"
        ],
        "summary": "Select next proxy",
        "description": "Select the next proxy without making an outbound target request.",
        "operationId": "rotate_proxy_api_rotate_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_ProxyResource_"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/proxies/stream": {
      "get": {
        "tags": [
          "Pool Management"
        ],
        "summary": "Stream all proxies as NDJSON",
        "description": "Stream all proxies in NDJSON format (newline-delimited JSON).\n\nUseful for exporting large proxy lists without buffering entire response.\nEach line is a complete JSON object representing one proxy.\n\nArgs:\n    status_filter: Filter by health status (optional)\n    rotator: ProxyWhirl dependency\n    api_key: API key verification\n\nReturns:\n    StreamingResponse with NDJSON content",
        "operationId": "stream_proxies_api_proxies_stream_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "status_filter",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status Filter"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/proxies/export": {
      "get": {
        "tags": [
          "Pool Management"
        ],
        "summary": "Export all proxies as streaming JSON",
        "description": "Stream all proxies as newline-delimited JSON (NDJSON).\n\nThis endpoint is functionally equivalent to ``/api/proxies/stream``\nand is provided for discoverability under the ``/export`` path.\n\nArgs:\n    status_filter: Filter by health status (optional)\n    rotator: ProxyWhirl dependency\n    api_key: API key verification\n\nReturns:\n    StreamingResponse with NDJSON content",
        "operationId": "export_proxies_api_proxies_export_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "status_filter",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status Filter"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/proxies/test": {
      "post": {
        "tags": [
          "Pool Management"
        ],
        "summary": "Health check proxies (deprecated)",
        "description": "Run health checks on specified proxies.\n\n**DEPRECATED:** Use `/api/proxies/health-check` instead.\nThis endpoint is kept for backward compatibility and will be removed in a future version.\n\nArgs:\n    request_data: Optional list of proxy IDs to check\n    rotator: ProxyWhirl dependency\n    api_key: API key verification\n\nReturns:\n    List of health check results",
        "operationId": "health_check_proxies_deprecated_api_proxies_test_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HealthCheckRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_list_HealthCheckResult__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "deprecated": true,
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/proxies/{proxy_id}": {
      "get": {
        "tags": [
          "Pool Management"
        ],
        "summary": "Get proxy by ID",
        "description": "Get details of a specific proxy.\n\nArgs:\n    proxy_id: Proxy identifier\n    rotator: ProxyWhirl dependency\n    api_key: API key verification\n\nReturns:\n    Proxy resource",
        "operationId": "get_proxy_api_proxies__proxy_id__get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "proxy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Proxy Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_ProxyResource_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Pool Management"
        ],
        "summary": "Delete proxy",
        "description": "Remove a proxy from the pool.\n\nArgs:\n    proxy_id: Proxy identifier\n    rotator: ProxyWhirl dependency\n    storage: Optional storage dependency\n    api_key: API key verification",
        "operationId": "delete_proxy_api_proxies__proxy_id__delete",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "proxy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Proxy Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProxiedRequest": {
        "properties": {
          "url": {
            "type": "string",
            "maxLength": 2083,
            "minLength": 1,
            "format": "uri",
            "title": "Url",
            "description": "Target URL to fetch through proxy"
          },
          "method": {
            "type": "string",
            "enum": [
              "GET",
              "POST",
              "PUT",
              "DELETE",
              "PATCH",
              "HEAD",
              "OPTIONS"
            ],
            "title": "Method",
            "description": "HTTP method",
            "default": "GET"
          },
          "headers": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Headers",
            "description": "Request headers to pass through"
          },
          "body": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Body",
            "description": "Request body (for POST/PUT/PATCH)"
          },
          "timeout": {
            "type": "integer",
            "maximum": 300,
            "exclusiveMinimum": 0,
            "title": "Timeout",
            "description": "Request timeout in seconds",
            "default": 30
          }
        },
        "type": "object",
        "required": [
          "url"
        ],
        "title": "ProxiedRequest",
        "description": "Request to make an HTTP request through a proxy.",
        "example": {
          "headers": {
            "User-Agent": "ProxyWhirl/1.0"
          },
          "method": "GET",
          "timeout": 30,
          "url": "https://httpbin.org/ip"
        }
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  },
...
}